llhttp: build static library instead of shared library
restinio 0.7.2, which depends on llhttp, tries to link with the cmake
module "llhttp:llhttp_dynamic". However, llhttp produces
"llhttp:llhttp_shared".
Because of this mismatch, build instead the static library and cmake
module of llhttp which is correctly recognized by restinio.

Note that llhttp shared library packaging is also disabled. This is
because the current llhttp CMakeLists.txt fails when trying to package
both.